small refactor and fix tests

+++ atc/runtime/runtimetest/volume.go
@@ -54,7 +54,7 @@ func (v Volume) Source() string {
 	return fmt.Sprintf("%s-worker", v.Handle())
 }
 
-func (v Volume) StreamIn(ctx context.Context, path string, compression compression.Compression, limitInMB int, reader io.Reader) error {
+func (v Volume) StreamIn(ctx context.Context, path string, compression compression.Compression, limitInMB float64, reader io.Reader) error {
 	return v.Content.StreamIn(ctx, path, compression.Encoding(), limitInMB, reader)
 }
 
@@ -82,7 +82,7 @@ func (v Volume) DBVolume() db.CreatedVolume {
 	return v.DBVolume_
 }
 
-func (vc VolumeContent) StreamIn(ctx context.Context, path string, encoding baggageclaim.Encoding, _ int, tarStream io.Reader) error {
+func (vc VolumeContent) StreamIn(ctx context.Context, path string, encoding baggageclaim.Encoding, _ float64, tarStream io.Reader) error {
 	if encoding != baggageclaim.GzipEncoding {
 		return errors.New("only gzip is supported for runtimetest.VolumeContent")
 	}
