+++ worker/baggageclaim/baggageclaimfakes/fake_volume.go
@@ -106,13 +106,13 @@ type FakeVolume struct {
 	setPropertyReturnsOnCall map[int]struct {
 		result1 error
 	}
-	StreamInStub        func(context.Context, string, baggageclaim.Encoding, int, io.Reader) error
+	StreamInStub        func(context.Context, string, baggageclaim.Encoding, float64, io.Reader) error
 	streamInMutex       sync.RWMutex
 	streamInArgsForCall []struct {
 		arg1 context.Context
 		arg2 string
 		arg3 baggageclaim.Encoding
-		arg4 int
+		arg4 float64
 		arg5 io.Reader
 	}
 	streamInReturns struct {
@@ -639,14 +639,14 @@ func (fake *FakeVolume) SetPropertyReturnsOnCall(i int, result1 error) {
 	}{result1}
 }
 
-func (fake *FakeVolume) StreamIn(arg1 context.Context, arg2 string, arg3 baggageclaim.Encoding, arg4 int, arg5 io.Reader) error {
+func (fake *FakeVolume) StreamIn(arg1 context.Context, arg2 string, arg3 baggageclaim.Encoding, arg4 float64, arg5 io.Reader) error {
 	fake.streamInMutex.Lock()
 	ret, specificReturn := fake.streamInReturnsOnCall[len(fake.streamInArgsForCall)]
 	fake.streamInArgsForCall = append(fake.streamInArgsForCall, struct {
 		arg1 context.Context
 		arg2 string
 		arg3 baggageclaim.Encoding
-		arg4 int
+		arg4 float64
 		arg5 io.Reader
 	}{arg1, arg2, arg3, arg4, arg5})
 	stub := fake.StreamInStub
@@ -668,13 +668,13 @@ func (fake *FakeVolume) StreamInCallCount() int {
 	return len(fake.streamInArgsForCall)
 }
 
-func (fake *FakeVolume) StreamInCalls(stub func(context.Context, string, baggageclaim.Encoding, int, io.Reader) error) {
+func (fake *FakeVolume) StreamInCalls(stub func(context.Context, string, baggageclaim.Encoding, float64, io.Reader) error) {
 	fake.streamInMutex.Lock()
 	defer fake.streamInMutex.Unlock()
 	fake.StreamInStub = stub
 }
 
-func (fake *FakeVolume) StreamInArgsForCall(i int) (context.Context, string, baggageclaim.Encoding, int, io.Reader) {
+func (fake *FakeVolume) StreamInArgsForCall(i int) (context.Context, string, baggageclaim.Encoding, float64, io.Reader) {
 	fake.streamInMutex.RLock()
 	defer fake.streamInMutex.RUnlock()
 	argsForCall := fake.streamInArgsForCall[i]
