+++ worker/baggageclaim/client/client.go
@@ -250,7 +250,7 @@ func (c *client) newVolume(apiVolume baggageclaim.VolumeResponse) baggageclaim.V
 	return volume
 }
 
-func (c *client) streamIn(ctx context.Context, destHandle string, path string, encoding baggageclaim.Encoding, limitInMB int, tarContent io.Reader) error {
+func (c *client) streamIn(ctx context.Context, destHandle string, path string, encoding baggageclaim.Encoding, limitInMB float64, tarContent io.Reader) error {
 	ctx, span := tracing.StartSpan(ctx, "volumeClient.streamIn", tracing.Attrs{
 		"volume":   destHandle,
 		"encoding": string(encoding),
@@ -263,7 +263,7 @@ func (c *client) streamIn(ctx context.Context, destHandle string, path string, e
 
 	request.URL.RawQuery = url.Values{
 		"path":  []string{path},
-		"limit": []string{fmt.Sprintf("%d", limitInMB)},
+		"limit": []string{fmt.Sprintf("%f", limitInMB)},
 	}.Encode()
 	if err != nil {
 		return err
